fixed const and size_t problems that were preventing make to run

+++ parsing/build_lst_utils.c
@@ -37,7 +37,7 @@ void	process_quotes(char *input, int *i, char *buffer, int *buf_index)
 }
 
 // Helper function to process escape sequences
-void	process_escape_sequence(char* input, int* i, char* buffer, int* buf_index)
+void	process_escape_sequence(char *input, size_t *i, char *buffer, int *buf_index)
 {
 	fprintf(stderr, "double slash if \n");
 	buffer[(*buf_index)++] = handle_escape_sequence(input, i, false);
@@ -66,7 +66,7 @@ void	process_env_variable(char* input, int* i, char* buffer, int* buf_index)
 }
 // Helper function to process regular text
 
-void	process_regular_text(char *input, int *i, char *buffer, int *buf_index)
+void	process_regular_text(char *input, size_t *i, char *buffer, int *buf_index)
 {
 	while (*i < ft_strlen(input) && !ft_isspace(input[*i])
 		&& input[*i] != '|' && input[*i] != '"'
